// src/types/event.ts
export interface Event {
  id: string;
  eventId?: string;
  title: string;
  start: Date;
  end: Date;
  startAt?: string;
  endAt?: string;
  description?: string;
  organizer?: string;
  source?: string;
  sourceId?: string;
  location?: string;
  attendees?: string[];
  calendar: 'google' | 'apple' | 'personal' | 'work';
  calendarName?: string;
  type?: 'service' | 'visit' | 'info' | 'task';
  leadId?: string;
  leadName?: string;
  leadSummary?: string;
  leadSource?: string;
  responsible?: string;
  team?: string;
  negotiationId?: string;
  color?: string;
  status?: 'pending' | 'completed' | 'overdue' | 'cancelled';
  cancelReason?: string;
  client?: string;
  phone?: string;
  notificationTime?: string;
  ownerName?: string;
  sourceType?: 'BillingRule' | 'Manual' | 'Integration';
  contractId?: string;
  invoiceId?: string;
  billingRuleId?: string;
  billingRuleStepId?: string;
  billingStatus?: 'pendente' | 'executado' | 'cancelado' | 'reagendado';
  billingLogs?: string[];
  backLink?: string;
}